refactor (card-item): - confusion between component and test file - update UT coverage (done)
