+++ src/app/home/components/card-item/card-item.component.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
-import { CardItemComponent } from './card-item.component.spec';
+import { CardItemComponent } from './card-item.component';
 
 @NgModule({
   declarations: [CardItemComponent],
